requirements.txt
streamlit
openai
faiss-cpu
pandas
polars
langchain
langchain-community
PyMuPDF
docx2txt
pytest
